# src/docs/swagger.yaml
definitions:
...
  github_com_itera-io_openstack-web-client_api_dto.GetComputeLimitResponse:
    properties:
      limits:
        $ref: '#/definitions/limits.Limits'
    type: object
...
  github_com_itera-io_openstack-web-client_api_dto.GetNetworkQuotaRequest:
    properties:
      projectId:
        type: string
    type: object
  github_com_itera-io_openstack-web-client_api_dto.GetNetworkQuotaResponse:
    properties:
      quota:
        $ref: '#/definitions/quotas.Quota'
    type: object
...
  github_com_itera-io_openstack-web-client_api_helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: '#/definitions/github_com_itera-io_openstack-web-client_api_helper.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/github_com_itera-io_openstack-web-client_api_validations.ValidationError'
        type: array
    type: object
  github_com_itera-io_openstack-web-client_api_helper.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 42901
    - 42902
    - 50001
    - 50002
    type: integer
    x-enum-varnames:
    - Success
    - ValidationError
    - AuthError
    - ForbiddenError
    - NotFoundError
    - LimiterError
    - OtpLimiterError
    - CustomRecovery
    - InternalError
  github_com_itera-io_openstack-web-client_api_validations.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
  limits.Absolute:
    properties:
      maxImageMeta:
        description: MaxImageMeta is the amount of image metadata available to a tenant.
        type: integer
      maxPersonality:
        description: MaxPersonality is the amount of personality/files available to
          a tenant.
        type: integer
      maxPersonalitySize:
        description: MaxPersonalitySize is the personality file size available to
          a tenant.
        type: integer
      maxSecurityGroupRules:
        description: |-
          MaxSecurityGroupRules is the number of security group rules available to
          a tenant.
        type: integer
      maxSecurityGroups:
        description: MaxSecurityGroups is the number of security groups available
          to a tenant.
        type: integer
      maxServerGroupMembers:
        description: |-
          MaxServerGroupMembers is the number of server group members available
          to a tenant.
        type: integer
      maxServerGroups:
        description: MaxServerGroups is the number of server groups available to a
          tenant.
        type: integer
      maxServerMeta:
        description: MaxServerMeta is the amount of server metadata available to a
          tenant.
        type: integer
      maxTotalCores:
        description: MaxTotalCores is the number of cores available to a tenant.
        type: integer
      maxTotalFloatingIps:
        description: MaxTotalFloatingIps is the number of floating IPs available to
          a tenant.
        type: integer
      maxTotalInstances:
        description: MaxTotalInstances is the number of instances/servers available
          to a tenant.
        type: integer
      maxTotalKeypairs:
        description: MaxTotalKeypairs is the total keypairs available to a tenant.
        type: integer
      maxTotalRAMSize:
        description: |-
          MaxTotalRAMSize is the total amount of RAM available to a tenant measured
          in megabytes (MB).
        type: integer
      totalCoresUsed:
        description: TotalCoresUsed is the number of cores currently in use.
        type: integer
      totalFloatingIpsUsed:
        description: TotalFloatingIpsUsed is the number of floating IPs in use.
        type: integer
      totalInstancesUsed:
        description: TotalInstancesUsed is the number of instances/servers in use.
        type: integer
      totalRAMUsed:
        description: TotalRAMUsed is the total RAM/memory in use measured in megabytes
          (MB).
        type: integer
      totalSecurityGroupsUsed:
        description: TotalSecurityGroupsUsed is the total number of security groups
          in use.
        type: integer
      totalServerGroupsUsed:
        description: TotalServerGroupsUsed is the total number of server groups in
          use.
        type: integer
    type: object
  limits.Limits:
    properties:
      absolute:
        allOf:
        - $ref: '#/definitions/limits.Absolute'
        description: Absolute contains the limits and usage information.
    type: object
  quotas.Quota:
    properties:
      floatingip:
        description: FloatingIP represents a number of floating IPs. A "-1" value
          means no limit.
        type: integer
      network:
        description: Network represents a number of networks. A "-1" value means no
          limit.
        type: integer
      port:
        description: Port represents a number of ports. A "-1" value means no limit.
        type: integer
      rbac_policy:
        description: RBACPolicy represents a number of RBAC policies. A "-1" value
          means no limit.
        type: integer
      router:
        description: Router represents a number of routers. A "-1" value means no
          limit.
        type: integer
      security_group:
        description: SecurityGroup represents a number of security groups. A "-1"
          value means no limit.
        type: integer
      security_group_rule:
        description: SecurityGroupRule represents a number of security group rules.
          A "-1" value means no limit.
        type: integer
      subnet:
        description: Subnet represents a number of subnets. A "-1" value means no
          limit.
        type: integer
      subnetpool:
        description: SubnetPool represents a number of subnet pools. A "-1" value
          means no limit.
        type: integer
      trunk:
        description: Trunk represents a number of trunks. A "-1" value means no limit.
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /v2/compute/limits:
    get:
      consumes:
      - application/json
      description: Get Compute Limits
      produces:
      - application/json
      responses:
        "200":
          description: GetComputeLimits response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_itera-io_openstack-web-client_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_itera-io_openstack-web-client_api_dto.GetComputeLimitResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_itera-io_openstack-web-client_api_helper.BaseHttpResponse'
        "401":
          description: Unauthorized request
          schema:
            $ref: '#/definitions/github_com_itera-io_openstack-web-client_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get Compute Limits
      tags:
      - ComputeLimits
...
  /v2/networking/quotas:
    get:
      consumes:
      - application/json
      description: Get Network Quotas
      parameters:
      - description: GetNetworkQuota Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_itera-io_openstack-web-client_api_dto.GetNetworkQuotaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: GetNetworkQuota response
          schema:
            allOf:
            - $ref: '#/definitions/github_com_itera-io_openstack-web-client_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_itera-io_openstack-web-client_api_dto.GetNetworkQuotaResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_itera-io_openstack-web-client_api_helper.BaseHttpResponse'
        "401":
          description: Unauthorized request
          schema:
            $ref: '#/definitions/github_com_itera-io_openstack-web-client_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get Network Quotas
      tags:
      - NetworkQuotas
...
swagger: "2.0"